fix: 카테고리 페이지 구성 컴포넌트 수정
