--- src/types/groupNavType.ts
@@ -1,10 +1,0 @@
-import { largeCategoryType } from "@/types/largeCategoryType"
-
-
-export interface groupNavType{
-  group: largeCategoryType[]
-  gx: number
-  handleOpen: React.MouseEventHandler<HTMLLIElement>
-  selectedLCategory: number
-  isOpen: Boolean
-}
